Changed timers to use doubles instead of floats.

+++ main.c
@@ -188,7 +188,7 @@ void del_timer( char *name ); */
 /* Timers */
 TIMER *get_timers( );
 int get_timer( );
-TIMER *add_timer( const char *name, float delay, void (*cb)( TIMER *self ), int d0, int d1, int d2 );
+TIMER *add_timer( const char *name, double delay, void (*cb)( TIMER *self ), int d0, int d1, int d2 );
 void del_timer( char *name );
 void unlink_timer( TIMER *timer );
 void free_timer( TIMER *timer );
@@ -2687,7 +2687,7 @@ void check_timers( )
 
 
 
-TIMER *add_timer( const char *name, float delay, void (*cb)( TIMER *timer ),int d0, int d1, int d2 )
+TIMER *add_timer( const char *name, double delay, void (*cb)( TIMER *timer ),int d0, int d1, int d2 )
 {
    TIMER *t;
    struct timeval now;
